refactor: draw beui with winit and wgpu instead of egui
beui owned its node tree but borrowed egui for everything below it, so it
could never leave eframe behind. It now carries its own stack:

- geometry, color and input types replace the egui ones
- a freetype and harfbuzz text engine, shaping and rasterizing the way
  crates/tablet and text_block already do, with a galley cache
- a Painter that records a display list, and a Context that holds the
  fonts, the frame input and the frame output
- a wgpu renderer drawing every rectangle and glyph as one instanced
  rounded rect pass, with a shelf packed glyph atlas
- a winit runner behind beui::run, so the demo is no longer an eframe app

The three new renderer tests paint offscreen through a real device, which
covers the shader, the clipping and the glyph atlas.

+++ crates/beui/src/base/fill.rs
@@ -1,7 +1,9 @@
 use std::any::Any;
 use std::collections::HashMap;
 
-use egui::{Color32, Painter, Rect, Vec2};
+use crate::color::Color32;
+use crate::geometry::{Rect, Vec2};
+use crate::painter::Painter;
 
 use crate::document::Document;
 use crate::node::{Element, InteractInput, NodeId};
@@ -43,7 +45,7 @@ impl Element for FillNode {
     }
 
     fn paint(&self, doc: &Document, painter: &Painter, rects: &HashMap<NodeId, Rect>, rect: Rect) {
-        painter.rect_filled(rect, self.corner_radius, self.color);
+        painter.rect_filled(rect, f32::from(self.corner_radius), self.color);
         if let Some(child) = self.child {
             crate::paint::paint(doc, painter, rects, child);
         }
